+++ ext/qt6cr/include/qt6cr.h
@@ -1303,6 +1303,8 @@ qt6cr_color_t qt6cr_graphics_colorize_effect_color(qt6cr_handle_t handle);
 void qt6cr_graphics_colorize_effect_set_color(qt6cr_handle_t handle, qt6cr_color_t value);
 double qt6cr_graphics_colorize_effect_strength(qt6cr_handle_t handle);
 void qt6cr_graphics_colorize_effect_set_strength(qt6cr_handle_t handle, double value);
+void qt6cr_graphics_colorize_effect_on_color_changed(qt6cr_handle_t handle, qt6cr_void_callback_t callback, void *userdata);
+void qt6cr_graphics_colorize_effect_on_strength_changed(qt6cr_handle_t handle, qt6cr_double_callback_t callback, void *userdata);
 
 qt6cr_handle_t qt6cr_graphics_drop_shadow_effect_create(qt6cr_handle_t parent);
 double qt6cr_graphics_drop_shadow_effect_blur_radius(qt6cr_handle_t handle);